move storefront module to recipe folder
