+++ projects/storefrontlib/src/public_api.ts
@@ -7,6 +7,6 @@ export * from './cms-pages/index';
 export * from './cms-structure/index';
 export * from './layout/index';
 export * from './recipes/index';
+export * from './recipes/storefront.module';
 export * from './shared/index';
 export * from './storefront-config';
-export * from './storefront.module';
