+++ projects/storefrontlib/src/recipes/b2c-storefront.module.ts
@@ -3,9 +3,9 @@ import { ConfigModule, provideConfig } from '@spartacus/core';
 import { CmsLibModule } from '../cms-components/cms-lib.module';
 import { LayoutModule } from '../layout/layout.module';
 import { StorefrontConfig } from '../storefront-config';
-import { StorefrontModule } from '../storefront.module';
 import { b2cLayoutConfig } from './config/b2c-layout-config';
 import { defaultCmsContentConfig } from './config/static-cms-structure/default-cms-content.config';
+import { StorefrontModule } from './storefront.module';
 
 @NgModule({
   imports: [
