+++ projects/storefrontlib/src/recipes/storefront.module.ts
@@ -5,10 +5,10 @@ import {
   provideConfig,
   SmartEditModule,
 } from '@spartacus/core';
-import { ProductDetailsPageModule } from './cms-pages/product-details-page/product-details-page.module';
-import { ProductListingPageModule } from './cms-pages/product-listing-page/product-listing-page.module';
-import { StorefrontFoundationModule } from './recipes/storefront-foundation.module';
-import { StorefrontConfig } from './storefront-config';
+import { ProductDetailsPageModule } from '../cms-pages/product-details-page/product-details-page.module';
+import { ProductListingPageModule } from '../cms-pages/product-listing-page/product-listing-page.module';
+import { StorefrontConfig } from '../storefront-config';
+import { StorefrontFoundationModule } from './storefront-foundation.module';
 
 @NgModule({
   imports: [
